requirements.txt
openai>=1.5.0
langchain
chromadb
faiss-cpu
pillow
python-dotenv
